Add missing parameters in BitChoiceGroup (#9974)
