+++ src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Inputs/ChoiceGroup/BitChoiceGroupDemo.razor
@@ -10,6 +10,7 @@
               SecondaryNames="@(["Radio", "RadioButton", "RadioGroup", "RadioButtonGroup"])"
               Description="ChoiceGroup let people select a single option from two or more choices."
               Parameters="componentParameters"
+              SubEnums="componentSubEnums"
               SubClasses="componentSubClasses"
               GitHubUrl="Inputs/ChoiceGroup/BitChoiceGroup.razor"
               GitHubDemoUrl="Inputs/ChoiceGroup/BitChoiceGroupDemo.razor">
